+++ examples/Backtesting/custom_analytical.py
@@ -82,7 +82,7 @@ objective = SingleTargetObjective(target=NumericalTarget(name="Target", mode="MI
 # For details on recommender objects, we refer to [`recommenders`](./../Basics/recommenders.md).
 
 seq_greedy_EI_recommender = TwoPhaseMetaRecommender(
-    recommender=SequentialGreedyRecommender(acqf="qEI"),
+    recommender=SequentialGreedyRecommender(acquisition_function="qEI"),
 )
 random_recommender = TwoPhaseMetaRecommender(recommender=RandomRecommender())