+++ tests/test_deprecations.py
@@ -200,7 +200,7 @@ def test_deprecated_objective_config_deserialization():
 def test_deprecated_acqfs(acqf):
     """Using the deprecated acqf raises a warning."""
     with pytest.warns(DeprecationWarning):
-        SequentialGreedyRecommender(acqf=acqf)
+        SequentialGreedyRecommender(acquisition_function=acqf)
 
     with pytest.warns(DeprecationWarning):
         AcquisitionFunction.from_dict({"type": acqf})