+++ tests/test_iterations.py
@@ -24,7 +24,7 @@ from .conftest import run_iterations
 # Settings of the individual components to be tested
 ########################################################################################
 valid_acquisition_functions = get_args(
-    get_type_hints(BayesianRecommender.__init__)["acqf"]
+    get_type_hints(BayesianRecommender.__init__)["acquisition_function"]
 )
 valid_surrogate_models = [cls() for cls in get_available_surrogates()]
 valid_initial_recommenders = [cls() for cls in get_subclasses(NonPredictiveRecommender)]